sql_support

Module debug_tools

Functions

define_debug_functions

You can call this function to add all sql functions provided by this module to your connection. The list of supported functions is described below.

print_query

Print the entire contents of an arbitrary query. A common usage would be to pass SELECT * FROM table