Module sql_support::each_chunk

Functions

default_max_variable_number
Returns SQLITE_LIMIT_VARIABLE_NUMBER as read from an in-memory connection and cached. connection and cached. That means this will return the wrong value if it's set to a lower value for a connection using this will return the wrong thing, but doing so is rare enough that we explicitly don't support it (why would you want to lower this at runtime?).

each_chunk
Helper for the case where you have a &[impl ToSql] of arbitrary length, but need one of no more than the connection's MAX_VARIABLE_NUMBER (rather, default_max_variable_number()). This is useful when performing batched updates.

each_chunk_mapped
A version of each_chunk for the case when the conversion to to_sql requires an custom intermediate step. For example, you might want to grab a property off of an arrray of records

each_sized_chunk

each_sized_chunk_mapped
Utility to help perform batched updates, inserts, queries, etc. This is the low-level version of this utility which is wrapped by each_chunk and each_chunk_mapped, and it allows you to provide both the mapping function, and the chunk size.