sql_support
Function escape_string_for_pragma

pub fn escape_string_for_pragma(s: &str) -> String

In PRAGMA foo=‘bar’, 'bar' must be a constant string (it cannot be a bound parameter), so we need to escape manually. According to https://www.sqlite.org/faq.html, the only character that must be escaped is the single quote, which is escaped by placing two single quotes in a row.