Module sql_support::interrupt

Structs

SqlInterruptHandle
A Sync+Send type which can be used allow someone to interrupt an operation, even if it happens while rust code (and not SQL) is executing.

SqlInterruptScope
A helper that can be used to determine if an interrupt request has come in while the object lives. This is used to avoid a case where we aren't running any queries when the request to stop comes in, but we're still not done (for example, maybe we've run some of the autocomplete matchers, and are about to start running the others. If we rely solely on sqlite3_interrupt(), we'd miss the message that we should stop).