Module sql_support::repeat

Structs

RepeatDisplay
Helper type for printing repeated strings more efficiently. You should use repeat_display, or one of the repeat_sql_* helpers to construct it.

Functions

repeat_display
Construct a RepeatDisplay that will repeatedly call fmt_one with a formatter count times, separated by sep.

repeat_multi_values
Returns a value that formats as num_values instances of (?,?,?,...) (where there are vars_per_value question marks separated by commas in between the ?s).

repeat_sql_values
Returns a value that formats as count instances of (?) separated by commas.

repeat_sql_vars
Returns a value that formats as count instances of ? separated by commas.