Struct sql_support::UncheckedTransaction

pub struct UncheckedTransaction<'conn> {
    pub conn: &'conn Connection,
    pub started_at: Instant,
    pub finished: bool,
}

rusqlite, in an attempt to save us from ourselves, needs a mutable ref to a connection to start a transaction. That is a bit of a PITA in some cases, so we offer this as an alternative - but the responsibility of ensuring there are no concurrent transactions is on our head.

This is very similar to the rusqlite Transaction - it doesn’t prevent against nested transactions but does allow you to use an immutable Connection.

FIXME: This currently won’t actually be used most of the time, because rusqlite added [Connection::unchecked_transaction] (and Transaction::new_unchecked, which can be used to reimplement unchecked_transaction_imm), which will be preferred in a call to c.unchecked_transaction(), because inherent methods have precedence over methods on extension traits. The exception here is that this will still be used by code which takes &impl ConnExt (I believe it would also be used if you attempted to call unchecked_transaction() on a non-Connection that implements ConnExt, such as a Safepoint, UncheckedTransaction, or Transaction itself, but such code is clearly broken, so is not worth considering).

The difference is that rusqlite’s version returns a normal rusqlite::Transaction, rather than the UncheckedTransaction from this crate. Aside from type’s name and location (and the fact that rusqlite’s detects slightly more misuse at compile time, and has more features), the main difference is: rusqlite’s does not track when a transaction began, which unfortunatly seems to be used by the coop-transaction management in places in some fashion.

There are at least two options for how to fix this:

Decide we don’t need this version, and delete it, and moving the transaction timing into the coop-transaction code directly (or something like this).

Decide this difference is important, and rename ConnExt::unchecked_transaction to something like ConnExt::transaction_unchecked.

Fields
conn: &'conn Connectionstarted_at: Instantfinished: bool
Implementations

impl<'conn> UncheckedTransaction<'conn>

pub fn new( conn: &'conn Connection, behavior: TransactionBehavior ) -> SqlResult<Self>

Begin a new unchecked transaction. Cannot be nested, but this is not enforced by Rust (hence ‘unchecked’) - however, it is enforced by SQLite; use a rusqlite savepoint for nested transactions.

pub fn commit(self) -> SqlResult<()>

Consumes and commits an unchecked transaction.

pub fn rollback(self) -> SqlResult<()>

Consumes and rolls back an unchecked transaction.