sql_support

Enum MaybeCached

pub enum MaybeCached<'conn> {
    Uncached(Statement<'conn>),
    Cached(CachedStatement<'conn>),
}

MaybeCached is a type that can be used to help abstract over cached and uncached rusqlite statements in a transparent manner.

Variants

Uncached(Statement<'conn>)

Cached(CachedStatement<'conn>)

Implementations

impl<'conn> MaybeCached<'conn>

pub fn prepare( conn: &'conn Connection, sql: &str, cached: bool, ) -> Result<MaybeCached<'conn>>