Struct sql_support::UncheckedTransaction

pub struct UncheckedTransaction<'conn> {
    pub conn: &'conn Connection,
    pub started_at: Instant,
    pub finished: bool,
}

rusqlite, in an attempt to save us from ourselves, needs a mutable ref to a connection to start a transaction. That is a bit of a PITA in some cases, so we offer this as an alternative - but the responsibility of ensuring there are no concurrent transactions is on our head.

This is very similar to the rusqlite Transaction - it doesn’t prevent against nested transactions but does allow you to use an immutable Connection.

Fields
conn: &'conn Connectionstarted_at: Instantfinished: bool
Implementations

impl<'conn> UncheckedTransaction<'conn>

pub fn new(
conn: &'conn Connection,
behavior: TransactionBehavior
) -> SqlResult<Self>

Begin a new unchecked transaction. Cannot be nested, but this is not enforced by Rust (hence ‘unchecked’) - however, it is enforced by SQLite; use a rusqlite savepoint for nested transactions.

pub fn commit(self) -> SqlResult<()>

Consumes and commits an unchecked transaction.

pub fn rollback(self) -> SqlResult<()>

Consumes and rolls back an unchecked transaction.